change icon and names of functions - editing projects
